% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_data}
\alias{check_input_data}
\title{Internal check: ensure required columns exist}
\usage{
check_input_data(data, need_ah = FALSE, need_odds = FALSE)
}
\description{
Internal check: ensure required columns exist
}
\keyword{internal}
