% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ah_outcome}
\alias{compute_ah_outcome}
\title{Compute Asian Handicap outcome (home team perspective)}
\usage{
compute_ah_outcome(hg, ag, handicap)
}
\description{
Compute Asian Handicap outcome (home team perspective)
}
\keyword{internal}
