% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_distribution.R
\name{odds_distribution}
\alias{odds_distribution}
\title{Odds distribution (initial skeleton)}
\usage{
odds_distribution(league, open_odds, data)
}
\description{
Odds distribution (initial skeleton)
}
