% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asian_handicap.R
\name{handicap_distribution}
\alias{handicap_distribution}
\title{Asian Handicap distribution (initial skeleton)}
\usage{
handicap_distribution(league, open_line, close_line, data)
}
\description{
Asian Handicap distribution (initial skeleton)
}
